# Do not add unnecessary R-labels (and an optimization)