--- a/Code/GraphMol/RGroupDecomposition/RGroupDecompData.h
+++ b/Code/GraphMol/RGroupDecomposition/RGroupDecompData.h
@@ -62,9 +62,7 @@ struct RGroupDecompData {
       RWMol *alignCore = core.first ? cores[0].core.get() : nullptr;
       CHECK_INVARIANT(params.prepareCore(*core.second.core, alignCore),
                       "Could not prepare at least one core");
-      if (params.onlyMatchAtRGroups) {
-        core.second.findIndicesWithRLabel();
-      }
+      core.second.findIndicesWithRLabel();
       core.second.countUserRGroups();
       core.second.labelledCore.reset(new RWMol(*core.second.core));
     }