--- a/ReleaseNotes.md
+++ b/ReleaseNotes.md
@@ -17,6 +17,13 @@
   (#3696)
 - There have been numerous changes to the RGroup Decomposition code which change
   the results. (#3767)
+- In RGroup Decomposition, when onlyMatchAtRGroups is set to false, each molecule
+  is now decomposed based on the first scaffold which adds/uses the least number
+  of non-user-provided R labels, rather than the first scaffold tout court.
+  Among other things, this allows the code to provide the same results for both
+  onlyMatchAtRGroups=true and onlyMatchAtRGroups=false when suitable scaffolds
+  are provided without requiring the user to get overly concerned about the
+  input ordering of the scaffolds. (#3969)
 - There have been numerous changes to `GenerateDepictionMatching2DStructure()` (#3811)
 - Setting the kekuleSmiles argument (doKekule in C++) to MolToSmiles will now
   cause the molecule to be kekulized before SMILES generation. Note that this