- use a `boost::dynamic_bitset` rather than a `std::set` for lookups
I replaced a `std::set` which was being used for lookups with a `dynamic_bitset` as it is more efficient. We consistently use `dynamic_bitset` for this purpose all over the RGD code, but apparently we missed this one.

- do not add unnecessary R-labels
While testing the RGD code on some internal datasets I ran into two cases that I believe were yielding suboptimal results when `onlyMatchAtRGroups=false`, because unnecessary R labels were being dynamically added.
Both are addressed by the new `testDoNotAddUnnecessaryRLabels()` and I will illustrate them below.

1. There is a single core
![image](https://user-images.githubusercontent.com/5244385/112223165-d630c400-8c29-11eb-8037-0c1c42a8c589.png)
while the molecules are
![image](https://user-images.githubusercontent.com/5244385/112223228-eb0d5780-8c29-11eb-9ff0-22e3b7c91057.png)
Before this PR, the code would not make a difference between assigning a H substituent vs a heavy atom substituent to user-defined or dynamically added labels.
The scoring function, when evaluating the various permutations, would then choose the permutation which adds the smallest number of labels, and so would choose the solution that yields only R1 and R2, with zero added labels.
However, when running with `NoSymmetrization`, in the situation above you end up with three labels (R1, R2, R3), since a single permutation is retained:

```
{'Core': ['c1ncc([*:3])c([*:2])c1[*:1]',
  'c1ncc([*:3])c([*:2])c1[*:1]',
  'c1ncc([*:3])c([*:2])c1[*:1]',
  'c1ncc([*:3])c([*:2])c1[*:1]'],
 'R1': ['F[*:1]', 'Cl[*:1]', 'O[*:1]', '[H][*:1]'],
 'R2': ['[H][*:2]', 'C[*:2]', '[H][*:2]', 'C[*:2]'],
 'R3': ['[H][*:3]', '[H][*:3]', '[H][*:3]', 'F[*:3]']}
```

There is a simple fix for this, which is filtering out all degenerate matches that would require adding a label whenever there is the possibility to choose an alternative degenerate match that does not require adding a label, because it accommodates all heavy substituents into existing labels, rather than assigning one of the existing labels to H and adding another label for the other heavy substituent. This fix has a negligible impact on performance.

2. There are two cores:
![image](https://user-images.githubusercontent.com/5244385/112225205-b8189300-8c2c-11eb-89dc-04ad359d7131.png)
while the molecules are
![image](https://user-images.githubusercontent.com/5244385/112225419-0f1e6800-8c2d-11eb-80ee-451b17890695.png)
Even though the first molecule matches the first core and does not require adding any label, before this PR both molecules would be associated to the second core:

```
{'Core': ['c1cc([*:1])c([*:3])cc1[*:4]', 'c1cc([*:1])c([*:3])cc1[*:4]'],
 'R1': ['Cl[*:1]', 'Br[*:1]'],
 'R3': ['C[*:3]', 'C[*:3]'],
 'R4': ['F[*:4]', '[H][*:4]']}
```

This happens because the first core already matches (with 2 additional labels), and the second core is never even tried, thus yielding a rather unexpected result, as labels should only be added when there is no better alternative.
After this PR, the expected result is generated:

```
{'Core': ['Fc1ccc([*:2])c([*:1])c1', 'c1ccc([*:3])c([*:1])c1'],
 'R1': ['C[*:1]', 'C[*:1]'],
 'R2': ['Cl[*:2]', ''],
 'R3': ['', 'Br[*:3]']}
```

where the first molecule is associated to the first core, and a single additional label is added to the second core to accommodate the second molecule.
I do realize that this fix requires probing all available cores rather than just picking the first that matches so is more expensive, but I believe getting the expected result is preferable to getting a suboptimal result, though a bit faster.