@greglandrum 2021-03-24
> This happens because the first core already matches (with 2 additional labels), and the second core is never even tried, thus yielding a rather unexpected result, as labels should only be added when there is no better alternative.
After this PR, the expected result is generated:
```
{'Core': ['Fc1ccc([*:2])c([*:1])c1', 'c1ccc([*:3])c([*:1])c1'],
 'R1': ['C[*:1]', 'C[*:1]'],
 'R2': ['Cl[*:2]', ''],
 'R3': ['', 'Br[*:3]']}
```
> where the first molecule is associated to the first core, and a single additional label is added to the second core to accommodate the second molecule.
> I do realize that this fix requires probing all available cores rather than just picking the first that matches so is more expensive, but I believe getting the expected result is preferable to getting a suboptimal result, though a bit faster.

Assuming I understand this correctly, I disagree: I think being able to say: "the cores are matched in the order provided and the code takes the first one that matches" is really helpful and makes the results more predictable. It's pretty straightforward for client code to reorder the cores to get the behavior you're looking for with this behavior change.