@greglandrum 2021-03-24
Fixing the behavior in your bullet point 1., on the other hand, makes sense to me.