@ptosco 2021-03-24
> Assuming I understand this correctly, I disagree: I think being able to say: "the cores are matched in the order provided and the code takes the first one that matches" is really helpful and makes the results more predictable. It's pretty straightforward for client code to reorder the cores to get the behavior you're looking for with this change.

@greglandrum Thanks for the review. I see the source of your concern, but my change does not impact on the core ordering - the order is still taken into account (and I do agree this is very important). Before the end of the morning I will come up with a better example (that I will include as a test case) that explains how the change works and how I think this improves the previous behavior.