@ptosco 2021-03-24
@greglandrum I have added a test to better illustrate my point and I will explain it below.
The algorithm still honors the ordering of cores. It will look for the first core in the order provided that matches the molecule without requiring adding labels. I think this is exactly the behavior that we want.
I have put together another example to illustrate my point.
Say that we have 3 cores, in this order:
![image](https://user-images.githubusercontent.com/5244385/112295612-5be95980-8c94-11eb-80e3-55cd8a54e244.png)

And these 5 molecules:
![image](https://user-images.githubusercontent.com/5244385/112295690-6dcafc80-8c94-11eb-99b4-3bbc6947b5a2.png)

Note that I have moved the core with fluorine at the first place, as you suggested.
The current `master` code will now yield these results, which I still think are suboptimal (I would be tempted to say "wrong"):
![image](https://user-images.githubusercontent.com/5244385/112296025-c7332b80-8c94-11eb-83e5-795e5a716dca.png)

The first molecule now correctly picks the first core, as it now the first choice.
However, the second molecule picks the second core, because it is the first that matches in the order provided. As a result, R3 and R4 are added.
To avoid this, I could have inverted the order of the second and third core:
![image](https://user-images.githubusercontent.com/5244385/112296515-49bbeb00-8c95-11eb-9cff-b86caeda8116.png)

but this would cause the third molecule to also pick the disubstituted core, while the monosubstituted core, being last, would never be picked:
![image](https://user-images.githubusercontent.com/5244385/112296719-60fad880-8c95-11eb-8a67-991ecfece659.png)

This means that I cannot achieve an optimal RGD by simply reordering the cores.

Instead, with the code in my PR, I can use this core ordering
![image](https://user-images.githubusercontent.com/5244385/112295612-5be95980-8c94-11eb-80e3-55cd8a54e244.png)

and get the results that I expect, with a single added R label rather than two, which are unnecessary:
![image](https://user-images.githubusercontent.com/5244385/112297038-b0d99f80-8c95-11eb-97af-1cedc7d6b4f7.png)

In summary, this is how the new algorithm works:
* cores are tried in the provided order
* for each core, the number of additional labels required to accommodate the molecule is recorded
* as soon as a core that requires no additional labels to accommodate the molecule is found, this core is accepted as no better core can be found
* After all cores are tried, the first core in the provided order that requires the smallest number of additional R labels is picked

I hope this make sense to you and you agree this is the most desirable behavior.