@ptosco 2021-03-24
Here's a couple of benchmarks I ran on an internal dataset with 25 cores and 7805 molecules.
The current PR speeds up RGD with `onlyMatchAtRGroups=True` by a factor >2. I believe most of the gain comes from early pruning of unnecessary symmetries (point 1 above).

### Current `master` timings with `onlyMatchAtRGroups=True`
```
%%timeit
params = RGroupDecompositionParameters()
params.onlyMatchAtRGroups = True
params.alignment = RGroupCoreAlignment.NoAlignment
RDLogger.DisableLog("rdApp.warning")
rgd = RGroupDecomposition(cores_r_groups, params)
n_accepted = 0
for m in mols:
    if (rgd.Add(m) != -1):
        n_accepted += 1
rgd.Process()
RDLogger.EnableLog("rdApp.warning")
print("n_accepted", n_accepted)

n_accepted 4818
20.3 s ± 486 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### Current `master` timings with `onlyMatchAtRGroups=False`
```
%%time
params = RGroupDecompositionParameters()
params.onlyMatchAtRGroups = False
params.alignment = RGroupCoreAlignment.NoAlignment
RDLogger.DisableLog("rdApp.warning")
rgd = RGroupDecomposition(cores_r_groups, params)
n_accepted = 0
for m in mols:
    if (rgd.Add(m) != -1):
        n_accepted += 1
rgd.Process()
RDLogger.EnableLog("rdApp.warning")
print("n_accepted", n_accepted)

n_accepted 5062
1min 59s ± 1.43 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### This PR timings with `onlyMatchAtRGroups=True`
```
%%timeit
params = RGroupDecompositionParameters()
params.onlyMatchAtRGroups = True
params.alignment = RGroupCoreAlignment.NoAlignment
RDLogger.DisableLog("rdApp.warning")
rgd = RGroupDecomposition(cores_r_groups, params)
n_accepted = 0
for m in mols:
    if (rgd.Add(m) != -1):
        n_accepted += 1
rgd.Process()
RDLogger.EnableLog("rdApp.warning")
print("n_accepted", n_accepted)

n_accepted 4818
21.2 s ± 253 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### This PR timings with `onlyMatchAtRGroups=False`
```
%%time
params = RGroupDecompositionParameters()
params.onlyMatchAtRGroups = False
params.alignment = RGroupCoreAlignment.NoAlignment
RDLogger.DisableLog("rdApp.warning")
rgd = RGroupDecomposition(cores_r_groups, params)
n_accepted = 0
for m in mols:
    if (rgd.Add(m) != -1):
        n_accepted += 1
rgd.Process()
RDLogger.EnableLog("rdApp.warning")
print("n_accepted", n_accepted)

n_accepted 5062
48.5 s ± 1.31 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```