@greglandrum 2021-03-24
Ok, Paolo and I just talked offline and I think I understand the goal here. I would express it a bit differently than Paolo did, so let me try that and he can check whether or not he agrees with my statement.

The current behavior (on master and in the previous releases) of the RGD code is to use the first scaffold which matches a molecule. The second part of this PR changes that so that the code uses the first scaffold which adds/uses the least number of  non-user-provided R groups. Among other things, this allows the code to provide the same results for both `onlyMatchAtRGroups =true` and `onlyMatchAtRGroups=false` when suitable scaffolds are present without requiring the user to get overly concerned about the ordering of the scaffolds.